# src/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "test API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "API functions",
      "description": "API functions of our application"
    }
  ],
  "paths": {
    "/users/{id}": {
      "get": {
        "summary": "returns simple answer from get",
        "tags": [
          "API functions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "simple parameter",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "description": "parameter id just for test",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "document": {
                "schema": {
                  "type": "string",
                  "example": "some text"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "returns json object from post",
        "tags": [
          "API functions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              },
              "example": {
                "country": "Russia",
                "city": "Spb"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "simple parameter",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "description": "parameter id just for test",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      }
    }
  }
}